refactor(storage): extract session tag/work-event/phase CRUD (polylogue-1r9c)
Problem: polylogue-1r9c asks for a source-grounded hotspot map of
Polylogue's largest execution control centers plus at least one
coherent extraction slice that makes a named control center materially
smaller with no duplicate execution path, filing the rest as durable
child beads.

What changed:

1. docs/architecture-hotspots.md — new hotspot map covering all eight
   named control centers (storage read tier 11.3k lines, API facade
   5.9k, storage repair 5.6k, daemon HTTP 4.6k, storage write tier
   4.6k, CLI query dispatch's _execute_archive_query_stdout 632-line
   function, daemon service loop's run_daemon_services 475-line
   function, MCP register_read_tools/register_mutation_tools) with
   file:line evidence, call-boundary/ownership-seam analysis, and a
   prioritized extraction sequence. Documents why this slice targeted
   the write tier rather than the two larger hotspots (storage repair's
   three internal blocks share an untraced pool of generic helpers and
   one dataclass whose declaration/usage location is already
   mismatched — building that dependency graph is real, separate work;
   the read tier has the widest fan-in in the codebase and needs a
   broader regression net than one pass's budget covers).

2. polylogue/storage/sqlite/archive_tiers/session_annotations_write.py
   (new) — the ArchiveSessionTag/ArchiveSessionWorkEvent/
   ArchiveSessionPhase dataclasses and their upsert_session_tag/
   read_session_tags/upsert_session_work_event/read_session_work_events/
   upsert_session_phase/read_session_phases functions, moved verbatim
   from write.py, plus the four small helpers only they use. This block
   had zero callers into or from write_parsed_session_to_archive or any
   other write.py function — confirmed by dependency tracing before
   moving anything, matching the AC's "no duplicate execution path"
   requirement. The one shared dependency (_json_dumps, used by 8 other
   write.py call sites so it correctly stays there) is imported back
   lazily inside the moved functions' bodies to avoid a module-init-time
   circular import — the same lazy-import pattern write.py's own
   _mirror_session_tag_assertion_if_available already used for its
   user_write.py dependency.

3. write.py re-exports all nine public names unchanged (same import
   path for every external caller: archive_tiers/archive.py and the
   test suite import ArchiveSessionPhase/ArchiveSessionWorkEvent from
   polylogue.storage.sqlite.archive_tiers.write exactly as before).
   write.py: 4,595 -> 4,210 lines (-385, -8.4%).

4. docs/plans/layering.yaml — the writer-module layering policy tracks
   which physical file owns which mutating entrypoint per tier by
   parsing each file's "Writer module: <tiers>." docstring marker
   against a manifest. Moved the session-tag-assertion-mirror twin-write
   contract's module reference and the three relocated entrypoints from
   write.py's manifest entry to a new session_annotations_write.py
   entry; write.py's own declared tiers narrow from [index, user] to
   [index] since none of its remaining entrypoints touch the user tier
   directly (verified via the checker's own _mutation_tiers analysis).
   This surfaced a separate, pre-existing imprecision in
   archive_tiers/archive.py's "raw-revision-authority" twin-write
   contract: the checker's transitive tier attribution is coarse
   (crediting a WHOLE imported module's declared surface for any call
   into it, not the specific called function's real footprint), so
   archive.py's claimed "user" tier reach was riding entirely on that
   coarseness through its write.py import — with write.py's declared
   surface correctly narrowed, static tracing of all 11 of archive.py's
   declared entrypoints confirms none of them touch the user tier
   through any traceable path. Corrected archive.py's docstring/manifest
   to [index, source] (dropping the never-actually-proven "user" claim)
   and replaced the raw-revision-authority contract with a
   raw-membership-classification contract whose surfaces match reality
   ([index, source], proven by apply_raw_membership_classification's
   traced footprint) — a documentation/proof-declaration correction with
   no runtime behavior change, not a new claim about what the code does.

5. devtools/docs_surface.py + docs/README.md + CLAUDE.md — index the
   new architecture-hotspots.md doc (the docs-surface render gate
   rejects unindexed documentation files) and add it to CLAUDE.md's
   reference-docs table.

6. Filed six child beads for the remaining seven control centers per
   AC-5 (each with an explicit non-goal so it can't balloon into a
   second decompose-everything mega-bead): polylogue-redt (storage read
   tier — build the helper-dependency graph first), polylogue-u5dw
   (storage repair — resolve the three blocks' shared-helper coupling +
   the c9y storage-vs-maintenance placement question), polylogue-1vzf
   (CLI query dispatch — registry-ize the output-format branches,
   following this bead's own write-effects-registry precedent),
   polylogue-gikp (API facade — investigate verb-group mixins),
   polylogue-kchb (daemon HTTP — extend the existing web_shell_*.py
   satellite pattern), polylogue-avmq (daemon service loop — blocked-by
   polylogue-yp0 since the event-bus work will reshape this function's
   body directly), polylogue-w9di (MCP tool registration — lowest
   priority, already has a delegation pattern in place).

Verification: mypy --strict clean on write.py + session_annotations_write.py
+ archive.py. ruff check clean. devtools verify --quick green end to
end (ruff/mypy/render/topology/layering/closure-matrix/manifests all
pass — layering specifically required the archive.py contract fix
above to go green, verified via a standalone script replicating the
checker's own _entrypoint_tiers/_mutation_tiers functions before and
after). devtools test tests/unit/storage/test_archive_tiers_write.py
tests/unit/storage/test_archive_tiers_self_verify.py (64 passed) and
tests/unit/storage/test_revision_replay.py (16 passed, covers
apply_raw_membership_classification specifically) — full behavioral
parity, zero regressions. devtools render all --check green (no "out
of sync"/"unindexed" findings).

Ref polylogue-1r9c.
